% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-seqone-gi-csv.R
\name{bind_seqone_gi_csv}
\alias{bind_seqone_gi_csv}
\title{Bind together SeqOne genomic instability information from csv files}
\usage{
bind_seqone_gi_csv(filepaths, folder, gi_csv_regex = "hrd-results.*csv")
}
\arguments{
\item{filepaths}{The full filepaths for the genomic instability csv files
to bind, which can be found using \code{list_seqone_gi_csv}.}

\item{folder}{The folder to copy files to.}

\item{gi_csv_regex}{The naming pattern for SeqOne GI files. Defaults
to "hrd-results.*csv"}
}
\value{
A dataframe of SeqOne GI results collated using \code{read_seqone_gi_csv}
}
\description{
This function can be used with \code{list_seqone_gi_csv} to search the S drive
for all SeqOne genomic instability csv files, and then collate them
together.
}
