% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-column.R
\name{check_column}
\alias{check_column}
\title{Check column input for collect_*_by_vector functions}
\usage{
check_column(query_col, output_cols, query_tbl)
}
\arguments{
\item{query_col}{The query_col argument of the collect_*_by_vector function}

\item{output_cols}{The output_cols argument of the collect_*_by_vector function}

\item{query_tbl}{The DNA Database table queried by the collect_*_by_vector function}
}
\value{
Error if column input is incorrect
}
\description{
Check column input for collect_*_by_vector functions
}
\examples{
collect_results_by_vector(query_col = "labno", query_vector = c("24024388"))
}
