% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-output.R
\name{check_output}
\alias{check_output}
\title{Check outpu of collect_*_by_vector functions}
\usage{
check_output(output)
}
\arguments{
\item{output}{The dataframe returned by the collect_*_by_vector function}
}
\value{
Errors and warnings if output dataframe is not correct format
}
\description{
Check outpu of collect_*_by_vector functions
}
\examples{
collect_results_by_vector(query_col = "labno", query_vector = c(""))
}
