% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-dnadb-ids-by-labno.R
\name{collect_dnadb_ids_by_labno}
\alias{collect_dnadb_ids_by_labno}
\title{Collect identifiers for samples from DNA Database by DNA lab number}
\usage{
collect_dnadb_ids_by_labno(
  labno_vector,
  cols = c("labno", "i_gene_r_no", "i_gene_s_no", "firstname", "surname", "nhsno",
    "pathno")
)
}
\arguments{
\item{labno_vector}{A character vector of DNA Database lab numbers}

\item{cols}{The columns to include from the DNA Database "Samples" table}
}
\value{
A dataframe of sample identifiers
}
\description{
Collect identifiers for samples from DNA Database by DNA lab number
}
\examples{
collect_dnadb_ids_by_labno(labno_vector = c("24039973"))
}
