% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-dnadb-results-by-labno.R
\name{collect_dnadb_results_by_labno}
\alias{collect_dnadb_results_by_labno}
\title{Collect results from DNA Database by DNA lab number}
\usage{
collect_dnadb_results_by_labno(
  labno_vector,
  cols = c("labno", "genodate", "resultsid", "test", "genotype", "genotype2", "genocomm")
)
}
\arguments{
\item{labno_vector}{A character vector of DNA Database lab numbers}

\item{cols}{The columns to include from the DNA Database "ResultsAccess" table}
}
\value{
A dataframe of sample results
}
\description{
Collect results from DNA Database by DNA lab number
}
\examples{
collect_dnadb_results_by_labno("24024388")
}
