% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-ids-by-vector.R
\name{collect_ids_by_vector}
\alias{collect_ids_by_vector}
\title{Collect DNA Database sample identifier information using a vector of search terms}
\usage{
collect_ids_by_vector(
  query_col,
  query_vector,
  output_cols = c("labno", "i_gene_r_no", "i_gene_s_no", "firstname", "surname", "nhsno",
    "pathno")
)
}
\arguments{
\item{query_col}{The column in the Samples table to search}

\item{query_vector}{A vector of search terms}

\item{output_cols}{The columns from the Samples table to return}
}
\value{
A dataframe of search results from the Samples table
}
\description{
Collect DNA Database sample identifier information using a vector of search terms
}
\examples{
collect_ids_by_vector(query_vector = c("24039973"), query_col = "labno")
}
