% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-results-by-vector.R
\name{collect_results_by_vector}
\alias{collect_results_by_vector}
\title{Collect DNA Database result information using a vector of search terms}
\usage{
collect_results_by_vector(
  query_col,
  query_vector,
  output_cols = c("labno", "genodate", "resultsid", "test", "genotype", "genotype2",
    "genocomm")
)
}
\arguments{
\item{query_col}{The column of the ResultsAccess table to search by}

\item{query_vector}{A vector of search terms}

\item{output_cols}{The columns of the ResultsAccess table to return}
}
\value{
A dataframe of search results from the ResultsAccess table
}
\description{
Collect DNA Database result information using a vector of search terms
}
\examples{
collect_results_by_vector(query_col = "labno", query_vector = c("24024388"))
}
