% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-seqone-gi-worksheets.R
\name{collect_seqone_gi_worksheets}
\alias{collect_seqone_gi_worksheets}
\title{Collect worksheets for SeqOne Genomic Instability testing from DNA Database}
\usage{
collect_seqone_gi_worksheets(
  seqone_strings = c("seqone", "seq\\\\sone", "seq_one", "SSXT\\\\ssWGS\\\\sHRD")
)
}
\arguments{
\item{seqone_strings}{A character vector of common strings used to describe
SeqOne Genomic Instability worksheets}
}
\value{
A dataframe returned from \code{collect_worksheets_by_description}.
}
\description{
Collect worksheets for SeqOne Genomic Instability testing from DNA Database
}
\examples{
gi_ws_info <- collect_seqone_gi_worksheets()
}
