% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-worksheets-by-vector.R
\name{collect_worksheets_by_vector}
\alias{collect_worksheets_by_vector}
\title{Collect DNA Database worksheet information using a vector of search terms}
\usage{
collect_worksheets_by_vector(
  query_col,
  query_vector,
  output_cols = c("pcrid", "date", "description")
)
}
\arguments{
\item{query_col}{The column in the PCR_New table to search by}

\item{query_vector}{A vector of search terms}

\item{output_cols}{The columns from the PCR_New table to return}
}
\value{
A dataframe of search results from the PCR_New table
}
\description{
Collect DNA Database worksheet information using a vector of search terms
}
\examples{
collect_worksheets_by_vector(query_col = "pcrid", query_vector = c(146462))
}
