% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-to-eval-hrd.R
\name{connect_to_eval_hrd}
\alias{connect_to_eval_hrd}
\title{Connect to the Eval_HRD table}
\usage{
connect_to_eval_hrd(
  catalog = "MolecularDB",
  schema = "GenExport",
  table = "Eval_HRD",
  col_vector = c("test_order_date", "test_name", "test_identifier", "referral_number",
    "nhsn_umber", "test_name", "field", "data_value", "box"),
  test_names = c("PANEL: M2.5 - SeqOne HRD Status",
    "PANEL: R207.1 - Inherited ovarian cancer (without breast cancer) v4.0 (ICP)",
    "PANEL: M2_tBRCA_PS", "PANEL: M2_tBRCA_PS v2.a", "PANEL: M2_tBRCA_PS v2.b")
)
}
\arguments{
\item{catalog}{The catalog to select the lazy table from}

\item{schema}{The schema to select the lazy table from}

\item{table}{The name of the table containing iGene HRD test results.}

\item{col_vector}{Character vector of columns to select. This defaults to an
order which seems to work. For some reason test_order_date should be
selected first to avoid an "Invalid Descriptor Index" error}

\item{test_names}{The names of analysis panels which are expected within this
table}
}
\value{
The Eval_HRD table as a dataframe
}
\description{
This is a wrapper function to connect to test results for the HRD
(Homologous Recombination Deficiency) service which are exported from
iGene in the "Eval_HRD" table.
}
\examples{
eval_hrd <- connect_to_eval_hrd()
}
