% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-to-lazy-tbl.R
\name{connect_to_lazy_tbl}
\alias{connect_to_lazy_tbl}
\title{Connect to a lazy table in the Molecular DB database}
\usage{
connect_to_lazy_tbl(table, catalog = "MolecularDB", schema = "dbo")
}
\arguments{
\item{table}{The name of the table to load}

\item{catalog}{The catalog. Defaults to "MolecularDB"}

\item{schema}{The schema. Defaults to "dbo"}
}
\value{
A lazy table
}
\description{
Connect to a lazy table in the Molecular DB database
}
\examples{
dnadb_pcr_new <- connect_to_lazy_tbl("PCR_New")
}
