% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-to-sql_server.R
\name{connect_to_sql_server}
\alias{connect_to_sql_server}
\title{Connect to the Microsoft SQL Server}
\usage{
connect_to_sql_server(db = "moldb")
}
\arguments{
\item{db}{The database to connect to. Defaults to "moldb" which is the
mirror version of DNA Database.}
}
\value{
A DBI connection to the database
}
\description{
Connect to the Microsoft SQL Server
}
