% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions-icp.R
\name{descriptions_icp}
\alias{descriptions_icp}
\title{Descriptions of Inherited Cancer Panel (ICP) worksheets}
\usage{
descriptions_icp()
}
\value{
A vector of unique descriptions for ICP worksheets.
}
\description{
Descriptions of Inherited Cancer Panel (ICP) worksheets
}
\examples{
icp_strings <- descriptions_icp()
}
