% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions-pansolid.R
\name{descriptions_pansolid}
\alias{descriptions_pansolid}
\title{Descriptions of PanSolid next generation sequencing worksheets}
\usage{
descriptions_pansolid()
}
\value{
A vector of unique descriptions for PanSolid worksheets
}
\description{
The "descriptions" column on DNA Database is free-type, which means that
there are many different ways that PanSolid next generation sequencing
worksheets have been described.
}
\examples{
pansolid_strings <- descriptions_pansolid()
}
