% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions-seqone-gi.R
\name{descriptions_seqone_gi}
\alias{descriptions_seqone_gi}
\title{Descriptions of SeqOne Genomic Instability worksheets}
\usage{
descriptions_seqone_gi()
}
\value{
A vector of unique descriptions of SeqOne GI worksheets
}
\description{
Descriptions of SeqOne Genomic Instability worksheets
}
\examples{
seqone_strings <- descriptions_seqone_gi()
}
