% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dnadb-ids-by-labno.R
\name{get_dnadb_ids_by_labno}
\alias{get_dnadb_ids_by_labno}
\title{Get identifiers for samples from DNA Database by DNA lab number}
\usage{
get_dnadb_ids_by_labno(
  labno_vector,
  id_cols = c("labno", "i_gene_r_no", "i_gene_s_no", "firstname", "surname", "nhsno",
    "pathno")
)
}
\arguments{
\item{labno_vector}{A character vector of DNA Database lab numbers}

\item{id_cols}{The columns to include from the DNA Database "Samples" table}
}
\value{
A dataframe of sample identifiers
}
\description{
Get identifiers for samples from DNA Database by DNA lab number
}
\examples{
get_dnadb_ids_by_labno(labno_vector = c("24039973"))
}
