% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-seqone-gi-csv.R
\name{list_seqone_gi_csv}
\alias{list_seqone_gi_csv}
\title{Find all filepaths for SeqOne genomic instability csv files}
\usage{
list_seqone_gi_csv(gi_csv_regex = "hrd-results.*csv")
}
\arguments{
\item{gi_csv_regex}{The naming pattern for SeqOne GI files. Defaults
to "hrd-results.*csv"}
}
\value{
A list containing: a list of all full filepaths to SeqOne GI files;
and a dataframe of worksheet information collected from DNA Database
}
\description{
Find all filepaths for SeqOne genomic instability csv files
}
\examples{
gi_ws_filepaths <- list_seqone_gi_csv()[[1]]
}
