% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-worksheet-filepaths.R
\name{list_worksheet_filepaths}
\alias{list_worksheet_filepaths}
\title{List filepaths for files within worksheet folders}
\usage{
list_worksheet_filepaths(
  worksheet,
  path = "S:/central shared/Genetics/Repository/WorksheetAnalysedData/",
  pattern,
  recursive = TRUE,
  full_names = TRUE
)
}
\arguments{
\item{worksheet}{Worksheet. Format "WS123456"}

\item{path}{The filepath to the S drive "WorksheetAnalysedData" repository}

\item{pattern}{Regular expression of filename pattern}

\item{recursive}{To be used by \code{list.files}. \code{TRUE} or \code{FALSE.}}

\item{full_names}{To be used by \code{list.files}. \code{TRUE} or \code{FALSE.}}
}
\value{
A list of worksheet filepaths
}
\description{
List filepaths for files within worksheet folders
}
\examples{
list_worksheet_filepaths(worksheet = "WS157517", pattern = "hrd-results.*csv")
}
