% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-seqone-gi-df.R
\name{mutate_seqone_gi_df}
\alias{mutate_seqone_gi_df}
\title{Annotate genomic instability data from the SeqOne service}
\usage{
mutate_seqone_gi_df(
  df,
  non_patient_strings = c("Seraseq", "GenQA"),
  name_col = "surname",
  worksheet_col = "worksheet",
  labno_col = "labno",
  validation_ws = c("WS133557", "WS134687", "WS134928", "WS135001", "WS135498",
    "WS136827", "WS138201", "WS138439", "WS138627", "WS147582", "WS149085", "WS149086")
)
}
\arguments{
\item{df}{Dataframe of collated SeqOne genomic instability data}

\item{non_patient_strings}{Character vector of names to identify non
patient samples (External Quality Assessment samples, controls etc)}

\item{name_col}{The column to search within for non patient strings}

\item{worksheet_col}{The column in the dataframe containing worksheet number}

\item{labno_col}{The column in the dataframe with DNA laboratory number}

\item{validation_ws}{Character vector of worksheets used for validation of
the SeqOne genomic instability service.}
}
\value{
The original dataframe with additional columns
}
\description{
Annotate genomic instability data from the SeqOne service
}
