% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-wider-by-test.R
\name{pivot_wider_by_test}
\alias{pivot_wider_by_test}
\title{Pivot a dataframe by the test column}
\usage{
pivot_wider_by_test(
  df,
  test_vector,
  names_prefix,
  id_cols = c("referral_number", "nhsn_umber", "test_order_date", "test_identifier"),
  names_from = "box",
  values_from = "data_value"
)
}
\arguments{
\item{df}{The dataframe to pivot. This function is designed to be used with
the "Eval_*" series of tables exported from iGene to the SQL server.}

\item{test_vector}{A character vector of test names to pivot by.}

\item{names_prefix}{The prefix to use for naming columns}

\item{id_cols}{Supplied to the \code{id_cols} argument of \code{pivot_wider}}

\item{names_from}{Supplied to the \code{names_from} argument of \code{pivot_wider}}

\item{values_from}{Supplied to the \code{values_from} argument of \code{pivot_wider}}
}
\value{
A dataframe of test results
}
\description{
Pivot a dataframe by the test column
}
\examples{

eval_hrd <- connect_to_eval_hrd()

pivot_wider_by_test(df = eval_hrd,
test_vector = c("PANEL: R207.1 - Inherited ovarian cancer (without breast cancer) v4.0 (ICP)"), 
names_prefix = "glvar")
}
