% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqone_gi_csv.R
\name{read_seqone_gi_csv}
\alias{read_seqone_gi_csv}
\title{Read data from a SeqOne Genomic Instability comma separated value file}
\usage{
read_seqone_gi_csv(filepath, n_rows = 1)
}
\arguments{
\item{filepath}{The full filepath to the csv file}

\item{n_rows}{The number of rows to read, which defaults to 1.}
}
\value{
The contents of the file as a dataframe
}
\description{
Read data from a SeqOne Genomic Instability comma separated value file
}
