% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-worksheets-by-description.R
\name{collect_worksheets_by_description}
\alias{collect_worksheets_by_description}
\title{Collect worksheet information by the "description" column in DNA Database}
\usage{
collect_worksheets_by_description(
  descriptions,
  df_cols = c(pcrid, date, description)
)
}
\arguments{
\item{descriptions}{A comma-separated character vector of different possible
descriptions for a worksheet.}

\item{df_cols}{The columns of the PCR_New dataframe to return}
}
\value{
A dataframe of worksheets including worksheet
}
\description{
Collect worksheet information by the "description" column in DNA Database
}
\examples{
collect_worksheets_by_description(descriptions = c("seqone"))
}
